[FIX] 모든 카트 정보 갖고오기 메인뷰모델로 이동
